Add encoders for basic operations
