+++ app-backend/tool/src/main/scala/ast/codec/MapAlgebraOperationCodecs.scala
@@ -1,0 +1,79 @@
+package com.azavea.rf.tool.ast.codec
+
+import com.azavea.rf.tool.ast._
+
+import geotrellis.raster.render._
+import io.circe._
+import io.circe.optics.JsonPath._
+import io.circe.syntax._
+import io.circe.generic.auto._
+
+import scala.util.Try
+import java.security.InvalidParameterException
+
+trait MapAlgebraOperationCodecs {
+  implicit def mapAlgebraDecoder: Decoder[MapAlgebraAST]
+  implicit def mapAlgebraEncoder: Encoder[MapAlgebraAST]
+
+  implicit lazy val decodeOperations = Decoder.instance[MapAlgebraAST.Operation] { ma =>
+    ma._symbol match {
+      case Some("+") => ma.as[MapAlgebraAST.Addition]
+      case Some("-") => ma.as[MapAlgebraAST.Subtraction]
+      case Some("/") => ma.as[MapAlgebraAST.Division]
+      case Some("*") => ma.as[MapAlgebraAST.Multiplication]
+      case Some("mask") => ma.as[MapAlgebraAST.Multiplication]
+      case Some("reclassify") => ma.as[MapAlgebraAST.Multiplication]
+      case Some(unrecognized) =>
+        throw new InvalidParameterException(s"'$unrecognized' is not a recognized map algebra operation")
+      case None =>
+        throw new InvalidParameterException(s"Required 'apply' property not found on MapAlgebraAST operation")
+    }
+  }
+
+  implicit lazy val encodeOperations: Encoder[MapAlgebraAST.Operation] = new Encoder[MapAlgebraAST.Operation] {
+    final def apply(op: MapAlgebraAST.Operation): Json = op match {
+      case addition: MapAlgebraAST.Addition => addition.asJson
+      case subtraction: MapAlgebraAST.Subtraction => subtraction.asJson
+      case division: MapAlgebraAST.Division => division.asJson
+      case multiplication: MapAlgebraAST.Multiplication => multiplication.asJson
+      case _ => ???
+    }
+  }
+
+  implicit lazy val decodeAddition: Decoder[MapAlgebraAST.Addition] =
+    Decoder.forProduct3("args", "id", "label")(MapAlgebraAST.Addition.apply)
+
+  implicit lazy val encodeAddition: Encoder[MapAlgebraAST.Addition] =
+    Encoder.forProduct3("args", "id", "label")(op => (op.args, op.id, op.label))
+
+  implicit lazy val decodeSubtraction: Decoder[MapAlgebraAST.Subtraction] =
+    Decoder.forProduct3("args", "id", "label")(MapAlgebraAST.Subtraction.apply)
+
+  implicit lazy val encodeSubtraction: Encoder[MapAlgebraAST.Subtraction] =
+    Encoder.forProduct3("args", "id", "label")(op => (op.args, op.id, op.label))
+
+  implicit lazy val decodeDivision: Decoder[MapAlgebraAST.Division] =
+    Decoder.forProduct3("args", "id", "label")(MapAlgebraAST.Division.apply)
+
+  implicit lazy val encodeDivision: Encoder[MapAlgebraAST.Division] =
+    Encoder.forProduct3("args", "id", "label")(op => (op.args, op.id, op.label))
+
+  implicit lazy val decodeMultiplication: Decoder[MapAlgebraAST.Multiplication] =
+    Decoder.forProduct3("args", "id", "label")(MapAlgebraAST.Multiplication.apply)
+
+  implicit lazy val encodeMultiplication: Encoder[MapAlgebraAST.Multiplication] =
+    Encoder.forProduct3("args", "id", "label")(op => (op.args, op.id, op.label))
+
+  implicit lazy val decodeMasking: Decoder[MapAlgebraAST.Masking] =
+    Decoder.forProduct3("args", "id", "label")(MapAlgebraAST.Masking.apply)
+
+  implicit lazy val encodeMasking: Encoder[MapAlgebraAST.Masking] =
+    Encoder.forProduct3("args", "id", "label")(op => (op.args, op.id, op.label))
+
+  implicit lazy val decodeReclassification: Decoder[MapAlgebraAST.Reclassification] =
+    Decoder.forProduct4("args", "id", "label", "classBreaks")(MapAlgebraAST.Reclassification.apply _)
+
+  implicit lazy val encodeReclassification: Encoder[MapAlgebraAST.Reclassification] =
+    Encoder.forProduct3("args", "id", "label")(op => (op.args, op.id, op.label))
+
+}
