+++ app-backend/tool/src/main/scala/ast/codec/package.scala
@@ -1,0 +1,42 @@
+package com.azavea.rf.tool.ast
+
+import geotrellis.raster.render._
+import io.circe._
+import io.circe.optics.JsonPath._
+import io.circe._
+import io.circe.generic.auto._
+
+import scala.util.Try
+import java.security.InvalidParameterException
+
+package object codec {
+  // Codec necessary for interpreting JSON keys as Double
+  implicit val decodeKeyDouble: KeyDecoder[Double] = new KeyDecoder[Double] {
+    final def apply(key: String): Option[Double] = Try(key.toDouble).toOption
+  }
+  implicit val encodeKeyDouble: KeyEncoder[Double] = new KeyEncoder[Double] {
+    final def apply(key: Double): String = key.toString
+  }
+
+  implicit lazy val classBoundaryDecoder: Decoder[ClassBoundaryType] =
+    Decoder[String].map {
+      case "lessThan" => LessThan
+      case "lessThanOrEqualTo" => LessThanOrEqualTo
+      case "exact" => Exact
+      case "greaterThanOrEqualTo" => GreaterThanOrEqualTo
+      case "greaterThan" => GreaterThan
+      case unrecognized =>
+        throw new InvalidParameterException(s"'$unrecognized' is not a recognized ClassBoundaryType")
+    }
+
+  implicit lazy val classBoundaryEncoder: Encoder[ClassBoundaryType] =
+    Encoder.encodeString.contramap[ClassBoundaryType]({ cbType =>
+      cbType match {
+        case LessThan => "lessThan"
+        case LessThanOrEqualTo => "lessThanOrEqualTo"
+        case Exact => "exact"
+        case GreaterThanOrEqualTo => "greaterThanOrEqualTo"
+        case GreaterThan => "greaterThan"
+      }
+    })
+}
